add push perf to es more checks before install

+++ examples/es.py
@@ -10,7 +10,7 @@ from __future__ import unicode_literals
 from __future__ import division
 
 from fabric.api import settings as fabric_settings
-from fabric.context_managers import cd, hide, shell_env
+from fabric.context_managers import cd, hide
 from fabric.contrib import files as fabric_files
 from fabric.operations import sudo, run, put, get
 from fabric.state import env
@@ -22,8 +22,6 @@ from pyLibrary.maths.randoms import Random
 from pyLibrary.meta import use_settings
 from pyLibrary.strings import expand_template
 from pyLibrary.thread.threads import Lock
-from pyLibrary.times.dates import Date
-from pyLibrary.times.durations import MONTH, WEEK, MINUTE, HOUR, DAY
 from spot.instance_manager import InstanceManager
 
 
@@ -87,9 +85,9 @@ class ESSpot(InstanceManager):
     def _install_es(self, gigabytes):
         volumes = self.instance.markup.drives
 
-        if not fabric_files.exists("/home/ec2-user/temp"):
+        if not fabric_files.exists("/usr/local/elasticsearch"):
             with cd("/home/ec2-user/"):
-                run("mkdir temp")
+                run("mkdir -p temp")
 
             with cd("/home/ec2-user/temp"):
                 run('wget -c --no-cookies --no-check-certificate --header "Cookie: s_cc=true; s_nr=1425654197863; s_sq=%5B%5BB%5D%5D; oraclelicense=accept-securebackup-cookie; gpw_e24=http%3A%2F%2Fwww.oracle.com%2Ftechnetwork%2Fjava%2Fjavase%2Fdownloads%2Fjre8-downloads-2133155.html" "http://download.oracle.com/otn-pub/java/jdk/8u40-b25/jre-8u40-linux-x64.rpm" --output-document="jdk-8u5-linux-x64.rpm"')
@@ -153,24 +151,25 @@ class ESSpot(InstanceManager):
 
             #INCREASE NUMBER OF FILE HANDLES
             # sudo("sysctl -w fs.file-max=64000")
-
         # COPY CONFIG FILE TO ES DIR
-        yml = File("./examples/config/es_config.yml").read().replace("\r", "")
-        yml = expand_template(yml, {
-            "id": Random.hex(length=8),
-            "data_paths": ",".join("/data"+unicode(i+1) for i, _ in enumerate(volumes))
-        })
-        File("./results/temp/elasticsearch.yml").write(yml)
-        put("./results/temp/elasticsearch.yml", '/usr/local/elasticsearch/config/elasticsearch.yml', use_sudo=True)
+        if not fabric_files.exists("/usr/local/elasticsearch/config/elasticsearch.yml"):
+            yml = File("./examples/config/es_config.yml").read().replace("\r", "")
+            yml = expand_template(yml, {
+                "id": Random.hex(length=8),
+                "data_paths": ",".join("/data"+unicode(i+1) for i, _ in enumerate(volumes))
+            })
+            File("./results/temp/elasticsearch.yml").write(yml)
+            put("./results/temp/elasticsearch.yml", '/usr/local/elasticsearch/config/elasticsearch.yml', use_sudo=True)
 
         # FOR SOME REASON THE export COMMAND DOES NOT SEEM TO WORK
         # THIS SCRIPT SETS THE ES_MIN_MEM/ES_MAX_MEM EXPLICITLY
-        sh = File("./examples/config/es_run.sh").read().replace("\r", "")
-        sh = expand_template(sh, {"memory": unicode(int(gigabytes/2))})
-        File("./results/temp/elasticsearch.in.sh").write(sh)
-        with cd("/home/ec2-user"):
-            put("./results/temp/elasticsearch.in.sh", './temp/elasticsearch.in.sh', use_sudo=True)
-            sudo("cp -f ./temp/elasticsearch.in.sh /usr/local/elasticsearch/bin/elasticsearch.in.sh")
+        if not fabric_files.exists("/usr/local/elasticsearch/bin/elasticsearch.in.sh"):
+            sh = File("./examples/config/es_run.sh").read().replace("\r", "")
+            sh = expand_template(sh, {"memory": unicode(int(gigabytes/2))})
+            File("./results/temp/elasticsearch.in.sh").write(sh)
+            with cd("/home/ec2-user"):
+                put("./results/temp/elasticsearch.in.sh", './temp/elasticsearch.in.sh', use_sudo=True)
+                sudo("cp -f ./temp/elasticsearch.in.sh /usr/local/elasticsearch/bin/elasticsearch.in.sh")
 
     def _install_indexer(self):
         Log.note("Install indexer at {{instance_id}} ({{address}})", instance_id=self.instance.id, address=self.instance.ip_address)
